requirements.txt
av==14.4.0
black==25.1.0
caffe2==0.8.1
cityscapesscripts==2.2.4
cloudpickle==3.1.1
dagr.egg==info
docutils==0.21.2
fairscale==0.4.13
fvcore==0.1.5.post20221221
h5py==3.11.0
hdf5plugin==4.4.0
hydra-core==1.3.2
iopath==0.1.9
loguru==0.7.2
matplotlib==3.7.5
megengine==1.13.1
mmcv==2.2.0
mmdet==3.3.0
ninja==1.11.1.1
numba==0.58.1
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.1
onnx_simplifier==0.4.10
onnxruntime==1.22.0
opencv_python==4.10.0.82
openvino==2025.1.0
Pillow==9.4.0
Pillow==11.2.1
psutil==5.9.8
pycocotools==2.0.7
Pygments==2.19.1
pytorch_lightning==2.5.1.post0
PyYAML==6.0.1
PyYAML==6.0.2
recommonmark==0.7.1
scikit_learn==1.3.2
scipy==1.15.3
seaborn==0.13.2
setuptools==69.5.1
Shapely==2.1.1
Sphinx==8.2.3
sphinx_rtd_theme==3.0.2
tabulate==0.9.0
tensorrt==10.11.0.33
termcolor==3.1.0
thop==0.1.1.post2209072238
timm==1.0.15
torch==2.1.0+cu121
torch_cluster==1.6.3
torch_geometric==2.6.1
torch_scatter==2.1.2
torch_spline_conv==1.2.2
torchvision==0.16.0
tqdm==4.66.4
wandb==0.19.11
